rename abac/prp/common to abac/common
